+++ src/components/chart/index.ts
@@ -1,2 +1,6 @@
 export { WeightChart } from "./WeightChart";
 export { WeightChartContent } from "./WeightChartContent";
+export { AnalyticsSection } from "./AnalyticsSection";
+export { PerformanceChart } from "./PerformanceChart";
+export { RiskReturnChart } from "./RiskReturnChart";
+export { SectorChart } from "./SectorChart";